ServingAPI/MLconDocker_Flask/requirements.txt
flask == 2.1.0
requests
numpy
pandas
joblib
sklearn
tid-my-modelo-PEC2-CMCN1978==0.0.3